Began work on disabling torchbind.
