+++ openequivariance/implementations/TensorProductBase.py
@@ -441,117 +441,3 @@ class TensorProductBase:
 
     def calculate_flops_double_backward(self, batch_size: int) -> dict:
         raise NotImplementedError("This needs to be implemented in your class")
-
-    def setup_torch_custom_op(self):
-        if not extlib.TORCH_COMPILE:
-            self.setup_nocompile_ops()
-
-    def setup_nocompile_ops(self):
-        # ----------------- Forward pass -----------------
-        @torch.library.custom_op(
-            f"openequivariance::tp_forward{self.tp_id}",
-            mutates_args=(),
-            device_types="cuda",
-        )
-        def forward(
-            L1_in: torch.Tensor, L2_in: torch.Tensor, weights: torch.Tensor
-        ) -> torch.Tensor:
-            L1_in_c, L2_in_c, weights_c = (
-                L1_in.contiguous(),
-                L2_in.contiguous(),
-                weights.contiguous(),
-            )
-            L3_out = torch.empty(
-                (L1_in_c.shape[0], self.L3.dim), dtype=L1_in.dtype, device="cuda"
-            )
-            self.forward_raw(
-                L1_in_c.shape[0],
-                L1_in_c.data_ptr(),
-                L2_in_c.data_ptr(),
-                L3_out.data_ptr(),
-                weights_c.data_ptr(),
-            )
-            return L3_out
-
-        @forward.register_fake
-        def _(L1_in, L2_in, weights):
-            return L1_in.new_empty(L1_in.shape[0], self.L3.dim)
-
-        self.forward = forward
-
-        # ---------------- Backward pass -----------------
-        @torch.library.custom_op(
-            f"openequivariance::tp_grad_helper{self.tp_id}",
-            mutates_args=(),
-            device_types="cuda",
-        )
-        def backward_helper(
-            L1_in: torch.Tensor,
-            L2_in: torch.Tensor,
-            weights: torch.Tensor,
-            L3_grad: torch.Tensor,
-        ) -> typing.List[torch.Tensor]:
-            L1_grad = torch.empty_like(L1_in)
-            L2_grad = torch.empty_like(L2_in)
-            weights_grad = torch.empty_like(weights)
-
-            if self.config.shared_weights:
-                weights_grad[:] = 0.0
-
-            self.backward_raw(
-                L1_in.shape[0],
-                L1_in.contiguous().data_ptr(),
-                L1_grad.data_ptr(),
-                L2_in.contiguous().data_ptr(),
-                L2_grad.data_ptr(),
-                weights.contiguous().data_ptr(),
-                weights_grad.data_ptr(),
-                L3_grad.contiguous().data_ptr(),
-            )
-
-            return [L1_grad, L2_grad, weights_grad]
-
-        @backward_helper.register_fake
-        def _(L1_in, L2_in, weights, L3_grad):
-            return [
-                L1_in.new_empty(*L1_in.shape),
-                L2_in.new_empty(*L2_in.shape),
-                weights.new_empty(*weights.shape),
-            ]
-
-        def setup_context(ctx, inputs, output):
-            ctx.L1_in, ctx.L2_in, ctx.weights = inputs
-
-        def backward(ctx, grad_output):
-            result = backward_helper(ctx.L1_in, ctx.L2_in, ctx.weights, grad_output)
-            return result[0], result[1], result[2]
-
-        self.forward.register_autograd(backward, setup_context=setup_context)
-
-        def setup_context_double_backward(ctx, inputs, output):
-            ctx.L1_in, ctx.L2_in, ctx.weights, ctx.L3_grad = inputs
-
-        def double_backward(ctx, grad_output):
-            A, B, C, D = ctx.L1_in, ctx.L2_in, ctx.L3_grad, ctx.weights
-            E, F, G = grad_output[0], grad_output[1], grad_output[2]
-
-            op1 = backward_helper(E, F, D, C)
-            op2 = backward_helper(A, B, G, C)
-            op3 = forward(E, B, D)
-            op4 = backward_helper(
-                E, B, D, C
-            )  # op4 and op5 could be combined with op3 and op6
-            op5 = backward_helper(A, F, D, C)
-            op6 = forward(A, F, D)
-            op7 = forward(A, B, G)
-
-            return (
-                op1[0] + op2[0],
-                op1[1] + op2[1],
-                (op4[2] + op5[2]),
-                (op3 + op6 + op7),
-            )
-
-        backward_helper.register_autograd(
-            double_backward, setup_context=setup_context_double_backward
-        )
