+++ openequivariance/implementations/convolution/ConvolutionBase.py
@@ -768,295 +768,3 @@ class ConvolutionBase:
             result[name] = check_similiarity(name, to_check, ground_truth, thresh)
 
         return result
-
-    def setup_torch_module(self):
-        if not extlib.TORCH_COMPILE:
-            self.setup_nocompile_ops()
-
-    def setup_nocompile_ops(self):
-        """
-        Need two different functions depending on whether the
-        convolution is deterministic.
-        """
-        if not self.deterministic:
-
-            @torch.library.custom_op(
-                f"openequivariance::conv_forward{self.conv_id}",
-                mutates_args=(),
-                device_types="cuda",
-            )
-            def forward(
-                L1_in: torch.Tensor,
-                L2_in: torch.Tensor,
-                weights: torch.Tensor,
-                rows: torch.Tensor,
-                cols: torch.Tensor,
-            ) -> torch.Tensor:
-                L1_in_c, L2_in_c, weights_c = (
-                    L1_in.contiguous(),
-                    L2_in.contiguous(),
-                    weights.contiguous(),
-                )
-                L3_out = torch.zeros(
-                    (L1_in_c.shape[0], self.L3.dim), dtype=L1_in.dtype, device="cuda"
-                )
-
-                self.internal.exec_conv_rawptrs(
-                    L1_in_c.data_ptr(),
-                    L2_in_c.data_ptr(),
-                    weights_c.data_ptr(),
-                    L3_out.data_ptr(),
-                    rows.data_ptr(),
-                    cols.data_ptr(),
-                    cols.shape[0],
-                    L1_in.shape[0],
-                    self.workspace_ptr,
-                )
-
-                return L3_out
-
-            @forward.register_fake
-            def _(L1_in, L2_in, weights, rows, cols):
-                return L1_in.new_empty(L1_in.shape[0], self.L3.dim)
-
-            self.forward = forward
-
-            @torch.library.custom_op(
-                f"openequivariance::conv_backward{self.conv_id}",
-                mutates_args=(),
-                device_types="cuda",
-            )
-            def backward_helper(
-                L1_in: torch.Tensor,
-                L2_in: torch.Tensor,
-                weights: torch.Tensor,
-                L3_grad: torch.Tensor,
-                rows: torch.Tensor,
-                cols: torch.Tensor,
-            ) -> typing.List[torch.Tensor]:
-                L1_grad = torch.zeros_like(L1_in)
-                L2_grad = torch.empty_like(L2_in)
-                weights_grad = torch.empty_like(weights)
-
-                self.internal.backward_rawptrs(
-                    L1_in.contiguous().data_ptr(),
-                    L1_grad.data_ptr(),
-                    L2_in.contiguous().data_ptr(),
-                    L2_grad.data_ptr(),
-                    weights.contiguous().data_ptr(),
-                    weights_grad.data_ptr(),
-                    L3_grad.contiguous().data_ptr(),
-                    rows.data_ptr(),
-                    cols.data_ptr(),
-                    rows.shape[0],
-                    L1_in.shape[0],
-                    self.workspace_ptr,
-                    0,
-                )
-
-                return [L1_grad, L2_grad, weights_grad]
-
-            @backward_helper.register_fake
-            def _(L1_in, L2_in, weights, L3_grad, rows, cols):
-                return [
-                    L1_in.new_empty(*L1_in.shape),
-                    L2_in.new_empty(*L2_in.shape),
-                    weights.new_empty(*weights.shape),
-                ]
-
-            def setup_context(ctx, inputs, output):
-                ctx.L1_in, ctx.L2_in, ctx.weights, ctx.rows, ctx.cols = inputs
-
-            def backward(ctx, grad_output):
-                result = backward_helper(
-                    ctx.L1_in, ctx.L2_in, ctx.weights, grad_output, ctx.rows, ctx.cols
-                )
-                return result[0], result[1], result[2], None, None
-
-            self.forward.register_autograd(backward, setup_context=setup_context)
-
-            def setup_context_double_backward(ctx, inputs, output):
-                ctx.L1_in, ctx.L2_in, ctx.weights, ctx.L3_grad, ctx.rows, ctx.cols = (
-                    inputs
-                )
-
-            def double_backward(ctx, grad_output):
-                A, B, C, D = ctx.L1_in, ctx.L2_in, ctx.L3_grad, ctx.weights
-                E, F, G = grad_output[0], grad_output[1], grad_output[2]
-                rows, cols = ctx.rows, ctx.cols
-
-                op1 = backward_helper(E, F, D, C, rows, cols)
-                op2 = backward_helper(A, B, G, C, rows, cols)
-                op3 = forward(E, B, D, rows, cols)
-                op4 = backward_helper(
-                    E, B, D, C, rows, cols
-                )  # op4 and op5 could be combined with op3 and op6
-                op5 = backward_helper(A, F, D, C, rows, cols)
-                op6 = forward(A, F, D, rows, cols)
-                op7 = forward(A, B, G, rows, cols)
-
-                return (
-                    op1[0] + op2[0],
-                    op1[1] + op2[1],
-                    (op4[2] + op5[2]),
-                    (op3 + op6 + op7),
-                    None,
-                    None,
-                )
-
-            backward_helper.register_autograd(
-                double_backward, setup_context=setup_context_double_backward
-            )
-
-        else:
-
-            @torch.library.custom_op(
-                f"openequivariance::conv_forward{self.conv_id}",
-                mutates_args=(),
-                device_types="cuda",
-            )
-            def forward(
-                L1_in: torch.Tensor,
-                L2_in: torch.Tensor,
-                weights: torch.Tensor,
-                rows: torch.Tensor,
-                cols: torch.Tensor,
-                transpose_perm: torch.Tensor,
-            ) -> torch.Tensor:
-                L1_in_c, L2_in_c, weights_c = (
-                    L1_in.contiguous(),
-                    L2_in.contiguous(),
-                    weights.contiguous(),
-                )
-                L3_out = torch.zeros(
-                    (L1_in_c.shape[0], self.L3.dim), dtype=L1_in.dtype, device="cuda"
-                )
-
-                self.internal.exec_conv_rawptrs(
-                    L1_in_c.data_ptr(),
-                    L2_in_c.data_ptr(),
-                    weights_c.data_ptr(),
-                    L3_out.data_ptr(),
-                    rows.data_ptr(),
-                    cols.data_ptr(),
-                    rows.shape[0],
-                    L1_in.shape[0],
-                    self.workspace_ptr,
-                )
-
-                return L3_out
-
-            @forward.register_fake
-            def _(L1_in, L2_in, weights, rows, cols, transpose_perm):
-                return L1_in.new_empty(L1_in.shape[0], self.L3.dim)
-
-            self.forward = forward
-
-            @torch.library.custom_op(
-                f"openequivariance::conv_backward{self.conv_id}",
-                mutates_args=(),
-                device_types="cuda",
-            )
-            def backward_helper(
-                L1_in: torch.Tensor,
-                L2_in: torch.Tensor,
-                weights: torch.Tensor,
-                L3_grad: torch.Tensor,
-                rows: torch.Tensor,
-                cols: torch.Tensor,
-                transpose_perm: torch.Tensor,
-            ) -> typing.List[torch.Tensor]:
-                L1_grad = torch.zeros_like(L1_in)
-                L2_grad = torch.empty_like(L2_in)
-                weights_grad = torch.empty_like(weights)
-
-                if self.config.shared_weights:
-                    weights_grad[:] = 0.0
-
-                self.internal.backward_rawptrs(
-                    L1_in.contiguous().data_ptr(),
-                    L1_grad.data_ptr(),
-                    L2_in.contiguous().data_ptr(),
-                    L2_grad.data_ptr(),
-                    weights.contiguous().data_ptr(),
-                    weights_grad.data_ptr(),
-                    L3_grad.contiguous().data_ptr(),
-                    rows.data_ptr(),
-                    cols.data_ptr(),
-                    rows.shape[0],
-                    L1_in.shape[0],
-                    self.workspace_ptr,
-                    transpose_perm.data_ptr(),
-                )
-
-                return [L1_grad, L2_grad, weights_grad]
-
-            @backward_helper.register_fake
-            def _(L1_in, L2_in, weights, L3_grad, rows, cols, transpose_perm):
-                return [
-                    L1_in.new_empty(*L1_in.shape),
-                    L2_in.new_empty(*L2_in.shape),
-                    weights.new_empty(*weights.shape),
-                ]
-
-            def setup_context(ctx, inputs, output):
-                (
-                    ctx.L1_in,
-                    ctx.L2_in,
-                    ctx.weights,
-                    ctx.rows,
-                    ctx.cols,
-                    ctx.transpose_perm,
-                ) = inputs
-
-            def backward(ctx, grad_output):
-                result = backward_helper(
-                    ctx.L1_in,
-                    ctx.L2_in,
-                    ctx.weights,
-                    grad_output,
-                    ctx.rows,
-                    ctx.cols,
-                    ctx.transpose_perm,
-                )
-                return result[0], result[1], result[2], None, None, None
-
-            self.forward.register_autograd(backward, setup_context=setup_context)
-
-            def setup_context_double_backward(ctx, inputs, output):
-                (
-                    ctx.L1_in,
-                    ctx.L2_in,
-                    ctx.weights,
-                    ctx.L3_grad,
-                    ctx.rows,
-                    ctx.cols,
-                    ctx.transpose_perm,
-                ) = inputs
-
-            def double_backward(ctx, grad_output):
-                A, B, C, D = ctx.L1_in, ctx.L2_in, ctx.L3_grad, ctx.weights
-                E, F, G = grad_output[0], grad_output[1], grad_output[2]
-                rows, cols, transpose_perm = ctx.rows, ctx.cols, ctx.transpose_perm
-
-                op1 = backward_helper(E, F, D, C, rows, cols, transpose_perm)
-                op2 = backward_helper(A, B, G, C, rows, cols, transpose_perm)
-                op3 = forward(E, B, D, rows, cols, transpose_perm)
-                op4 = backward_helper(E, B, D, C, rows, cols, transpose_perm)
-                op5 = backward_helper(A, F, D, C, rows, cols, transpose_perm)
-                op6 = forward(A, F, D, rows, cols, transpose_perm)
-                op7 = forward(A, B, G, rows, cols, transpose_perm)
-
-                return (
-                    op1[0] + op2[0],
-                    op1[1] + op2[1],
-                    (op4[2] + op5[2]),
-                    (op3 + op6 + op7),
-                    None,
-                    None,
-                    None,
-                )
-
-            backward_helper.register_autograd(
-                double_backward, setup_context=setup_context_double_backward
-            )
